[rename] : CreditHistory, CreditType 패키지 이동
