+++ src/main/java/com/dnd/gongmuin/credit_history/repository/CreditHistoryRepository.java
@@ -2,7 +2,7 @@ package com.dnd.gongmuin.credit_history.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dnd.gongmuin.credit_history.CreditHistory;
+import com.dnd.gongmuin.credit_history.domain.CreditHistory;
 
 public interface CreditHistoryRepository extends JpaRepository<CreditHistory, Long> {
 }
